refactor of Asset
