+++ src/main/java/com/jala/search/controller/ControllerSearchAdvanced.java
@@ -17,9 +17,8 @@ import com.jala.search.models.SearchFile;
 import com.jala.search.models.TernaryBooleanEnum;
 import com.jala.utils.Logs;
 import com.jala.view.JPanelSearchAdvanced;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.log4j.Logger;
-import java.io.File;
+
 import java.util.List;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
