+++ src/main/java/com/jala/search/models/ISearchable.java
@@ -12,7 +12,6 @@
 
 package com.jala.search.models;
 
-import java.io.File;
 import java.util.List;
 
 /**
